backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "anthropic>=0.83.0",
    "celery>=5.6.2",
    "docker>=7.1.0",
    "fastapi[standard]>=0.129.0",
    "flask>=3.1.3",
    "gevent>=25.9.1",
    "langchain-anthropic>=1.3.3",
    "langgraph>=1.0.9",
    "litellm>=1.82.0",
    "openai>=2.21.0",
    "pinecone>=8.0.1",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "redis>=7.2.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-typescript>=0.23.2",
]
